# apps/docs/apis/openapi.yaml
openapi: 3.1.0
info:
  title: AI GNT API
  version: 0.1.0
paths:
  /v1/brain/summary:
    get:
      tags:
      - brain
      summary: Brain Summary
      operationId: brain_summary_v1_brain_summary_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /v1/onboarding/status:
    get:
      tags:
      - brain
      summary: Onboarding Status
      description: 'This org''s onboarding progress — "how close is this org to a
        working

        setup" — aggregated off onboarding_events (gnt/onboarding_metrics.py).

        Planned future ROI metering will read off the same table, so counts

        stay a flat event_type -> count map here rather than baking in today''s

        five known types.


        connected_cli isn''t an onboarding_event (gnt login has no event-log

        call site, unlike the Slack/GitHub connect routes and rule proposal) —

        it''s read straight off mcp_api_keys instead. A CLI key, once minted,

        always stays a row here even if later revoked, so this answers "has

        this org ever completed gnt login", not "is a CLI key live right now".

        apps/web/components/onboarding-status.tsx polls this for the /welcome

        page''s live step checklist.


        reached_five_rules_milestone is this product''s "org reaches N approved

        rules in first session" success metric, N=5 — derived straight off

        the same counts, not a second query,

        since rules_approved already is that lifetime-cumulative count (see

        onboarding_metrics.RULES_APPROVED_MILESTONE for why that''s an

        accurate proxy for "first session" here).'
      operationId: onboarding_status_v1_onboarding_status_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
...
components:
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: https://api.gntai.dev
  description: Production